+++ src/main/java/dk/eamv/ferrari/sharedcomponents/filter/forms/FormFactory.java
@@ -32,74 +32,6 @@ public class FormFactory {
         CUSTOMER
     }
 
-    private static int[] createFieldsString(GridPane gridPane, ArrayList<TextField> fieldsList, int column, int row, String... input) {
-        for (String i : input) {
-            VBox vBox = new VBox();
-            Label heading = new Label(i);
-            TextField textField = new TextField();
-            textField.setPromptText(i);
-            vBox.getChildren().addAll(heading, textField);
-            if (column > 2) {
-                column = 0;
-                row++;
-            }
-            gridPane.add(vBox, column, row);
-            fieldsList.add(textField);
-            column++;
-        }
-        return new int[] { column, row };
-    }
-
-    private static int[] createFieldsInt(GridPane gridPane, ArrayList<TextField> fieldsList, int column, int row, String... input) {
-        for (String i : input) {
-            VBox vBox = new VBox();
-            Label heading = new Label(i);
-            NumericTextField textField = new NumericTextField();
-            textField.setPromptText(i);
-            vBox.getChildren().addAll(heading, textField);
-            if (column > 2) {
-                column = 0;
-                row++;
-            }
-            gridPane.add(vBox, column, row);
-            fieldsList.add(textField);
-            column++;
-        }
-
-        return new int[] { column, row };
-    }
-    
-    private static GridPane createGridPane() {
-        GridPane gridPane = new GridPane();
-        gridPane.setVgap(25);
-        gridPane.setHgap(50);
-        gridPane.setAlignment(Pos.CENTER);
-
-        return gridPane;
-    }
-
-    private static Form createCustomerForm() {
-        Form customerForm = new Form(createGridPane(), new ArrayList<TextField>());
-        int[] fields = createFieldsString(customerForm.getGridPane(), customerForm.getFieldsList(), 0, 0, "Fornavn", "Efternavn", "Email", "Adresse");
-        createFieldsInt(customerForm.getGridPane(), customerForm.getFieldsList(), fields[0], fields[1], "Telefonnummer", "CPR");
-        return customerForm;
-    }
-
-    private static Form createCarForm() {
-        Form carForm = new Form(createGridPane(), new ArrayList<TextField>());
-        int[] fields = createFieldsInt(carForm.getGridPane(), carForm.getFieldsList(), 0, 0, "Årgang", "Pris", "Stelnummer");
-        createFieldsString(carForm.getGridPane(), carForm.getFieldsList(), fields[0], fields[1], "Model");
-        return carForm;
-    }
-    
-    private static Form createLoanForm() {
-        Form loanForm = new Form(createGridPane(), new ArrayList<TextField>());
-        createFieldsInt(loanForm.getGridPane(), loanForm.getFieldsList(), 0, 0, "Stelnummer", "Kunde CPR",
-                "Lånets størrelse", "Udbetaling", "Rente",
-                "Start dato", "Forfaldsdag");
-        return loanForm;
-    }
-
     private static Dialog wrap(Form form, Type table) {
         Dialog dialog = new Dialog<>();
